+++ carp_mobile_sensing/lib/domain/sampling_schema.dart
@@ -40,12 +40,12 @@ class SamplingSchema {
     List<SamplingConfiguration> configurationList = [];
 
     // since we're using json serialization below, make sure that the json
-    // functions have been registred
+    // functions have been registered
     CarpMobileSensing();
 
     for (var type in types) {
       if (configurations.containsKey(type)) {
-        // using json encoding/decoding to clone the measure object
+        // clone the measure object
         final configurationAsJson = _encode(configurations[type]);
         final SamplingConfiguration configurationClone =
             SamplingConfiguration.fromJson(